Crate icu_provider_fs

icu_fs_data_provider is one of the [ICU4X] components.

It reads ICU4X data files from the filesystem in a given directory. It can also export data to the filesystem via an iterable data provider (see the export module).

Examples

use icu_provider_fs::FsDataProvider;

let provider = FsDataProvider::try_new("/path/to/data/directory")
    .expect_err("Specify a real directoroy in the line above");

Modules

export
The export feature enables you to pull all data from some other data provider and persist it on the filesystem to be read by an FsDataProvider at runtime.

manifest

Structs

FsDataProvider
A data provider that reads ICU4X data from a filesystem directory.

Enums

FsDataError